feat: Add CRUD functionality for courses
This commit adds the necessary code to implement CRUD (Create, Read, Update, Delete) functionality for courses. It includes the creation of the cursoController.js file, which contains the routes and logic for handling course-related operations such as retrieving a list of courses, getting a course by ID, creating a new course, updating a course, and deleting a course.

The code also includes the necessary route handlers and logic for handling HTTP requests related to courses, such as GET requests to retrieve a list of courses or a specific course by ID, POST requests to create a new course, PUT requests to update a course, and DELETE requests to delete a course.

This commit is a feature addition and enhances the functionality of the application by allowing users to perform CRUD operations on courses.
