+++ index.js
@@ -16,17 +16,6 @@ const PORT = process.env.PORT || 3000;
 // Middleware para parsear el cuerpo de las solicitudes con formato JSON
 app.use(express.json());
 
-// Simulamos una base de datos con listas de estudiantes y cursos
-const estudiantes = [
-  { id: 1, nombre: "Juan Pérez", edad: 20 },
-  { id: 2, nombre: "Ana Gómez", edad: 22 },
-];
-
-const cursos = [
-  { id: 1, nombre: "Curso de Node.js", duracion: "4 semanas" },
-  { id: 2, nombre: "Curso de React", duracion: "6 semanas" },
-];
-
 // Creamos una ruta de ejemplo
 app.get("/", (req, res) => {
   res.sendFile(path.join(__dirname, "views", "index.html"));
@@ -117,88 +106,6 @@ app.delete("/estudiante/:id", (req, res) => {
 
   res.json({
     mensaje: `Estudiante con ID: ${id} eliminado exitosamente`,
-  });
-});
-
-// Ruta cursos
-app.get("/cursos", (req, res) => {
-  res.json({ mensaje: "Lista de cursos", cursos });
-});
-
-// Ruta cursos por ID
-app.get("/curso/:id", (req, res) => {
-  const { id } = req.params; // Extraemos el parámetro 'id' de la URL
-  const curso = cursos.find((e) => e.id === parseInt(id)); // Buscamos el curso por ID
-
-  if (!curso) {
-    // Si no se encuentra el curso, devolvemos un error 404
-    return res.status(404).json({ error: "Curso no encontrado" });
-  }
-
-  // Si se encuentra el curso, devolvemos la información
-  res.json({
-    mensaje: `Información del curso con ID: ${id}`,
-    curso: curso,
-  });
-});
-
-// Ruta para crear un nuevo curso
-app.post("/cursos", (req, res) => {
-  const { nombre, duracion } = req.body;
-
-  if (!nombre || !duracion) {
-    return res
-      .status(400)
-      .json({ error: "El nombre y la duración son requeridos" });
-  }
-
-  const nuevoCurso = {
-    id: cursos.length + 1,
-    nombre,
-    duracion,
-  };
-
-  cursos.push(nuevoCurso);
-
-  res.status(201).json({
-    mensaje: "Curso creado exitosamente",
-    curso: nuevoCurso,
-  });
-});
-
-// Ruta para actualizar un curso
-app.put("/curso/:id", (req, res) => {
-  const { id } = req.params;
-  const { nombre, duracion } = req.body;
-
-  const curso = cursos.find((c) => c.id === parseInt(id));
-
-  if (!curso) {
-    return res.status(404).json({ error: "Curso no encontrado" });
-  }
-
-  curso.nombre = nombre || curso.nombre;
-  curso.duracion = duracion || curso.duracion;
-
-  res.json({
-    mensaje: `Curso con ID: ${id} actualizado exitosamente`,
-    curso,
-  });
-});
-
-// Ruta para eliminar un curso
-app.delete("/curso/:id", (req, res) => {
-  const { id } = req.params;
-  const indice = cursos.findIndex((c) => c.id === parseInt(id));
-
-  if (indice === -1) {
-    return res.status(404).json({ error: "Curso no encontrado" });
-  }
-
-  cursos.splice(indice, 1);
-
-  res.json({
-    mensaje: `Curso con ID: ${id} eliminado exitosamente`,
   });
 });
